% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{build_network}
\alias{build_network}
\title{Description}
\usage{
build_network(data_crossed, data_by_structure, number_of_nodes = 60)
}
\arguments{
\item{data_crossed}{data_crossed (by author or affiliation)}

\item{data_by_structure}{data_authors}

\item{number_of_nodes}{the number of nodes}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
data_authors=tidy_authors(data)
data_crossed=cross(data_ref_authors,data_authors,var_to_cross="name")
graph=build_network(data_crossed,data_by_structure=data_authors,number_of_nodes=100)
}
