% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Description}
\usage{
cross(data_by_ref, data_by_structure, var_to_cross)
}
\arguments{
\item{data_by_ref}{data with one line=one ref}

\item{data_by_structure}{data with one line=one structure or person}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
data_authors=tidy_authors(data)
HALtere::cross(data_ref_authors,data_authors,var_to_cross="affiliation")
HALtere::cross(data_ref_authors,data_authors,var_to_cross="name")
}
