% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_by_group}
\alias{cross_by_group}
\title{Description}
\usage{
cross_by_group(data, var_to_cross = "name")
}
\arguments{
\item{data}{a tibble produced with the tidy_authors() function}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
cross_by_group(data_ref_authors \%>\% filter(id_ref==1),var_to_cross="affiliation")
}
