% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_collection.R
\name{extract_collection}
\alias{extract_collection}
\title{Description}
\usage{
extract_collection(collection, nmax = +Inf)
}
\arguments{
\item{collection}{name of a HAL collection}
}
\value{
a tibble
}
\description{
Description
}
\examples{
extract_collection("BIOEENVIS", nmax=20)
}
