% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_authors.R
\name{tidy_groups}
\alias{tidy_groups}
\title{Description}
\usage{
tidy_groups(data_ref_authors, method = "shortest", type = "people")
}
\arguments{
\item{data_ref_authors}{a tibble produced with the extract_collection() function}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("EVS_UMR5600", nmax=200)
data_ref_authors=tidy_ref_authors(data)
data_labs=tidy_groups(data_ref_authors,type="labs")
}
