% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_authors.R
\name{tidy_ref_authors}
\alias{tidy_ref_authors}
\title{Description}
\usage{
tidy_ref_authors(data, method = "shortest")
}
\arguments{
\item{data}{a tibble produced with the extract_collection() function}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
}
