% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{build_network}
\alias{build_network}
\title{Description}
\usage{
build_network(data_crossed, data_groups, number_of_nodes = 60)
}
\arguments{
\item{data_crossed}{data_crossed (by author or affiliation)}

\item{data_groups}{data_authors}

\item{number_of_nodes}{the number of nodes}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)

data_groups=tidy_groups(data_ref_authors,type="people")
data_crossed=HALtere::cross(data_groups)
graph=build_network(data_crossed,data_groups,number_of_nodes=200)

data_groups=tidy_groups(data_ref_authors,type="labs")
data_crossed=HALtere::cross(data_groups)
graph=build_network(data_crossed,data_groups,number_of_nodes=200)

plot_network(graph, colorvar="name")
}
