% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Description}
\usage{
cross(data_groups)
}
\arguments{
\item{data_groups}{data with one line=one structure or person}

\item{data_by_ref}{data with one line=one ref}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=200)
data_ref_authors=tidy_ref_authors(data)
data_groups=tidy_groups(data)
HALtere::cross(data_groups)
data_groups=tidy_groups(data_ref_authors,type="labs")
HALtere::cross(data_groups)
}
