% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_by_group}
\alias{cross_by_group}
\title{Description}
\usage{
cross_by_group(data_ref_authors_oneref)
}
\arguments{
\item{data}{a tibble produced with the tidy_ref_authors() function and containing lines only for one reference}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data=extract_collection("BIOEENVIS", nmax=+Inf)
data_ref_authors=tidy_ref_authors(data)
data_groups=tidy_groups(data_ref_authors)
cross_by_group(data_ref_authors \%>\% dplyr::filter(id_ref==1))
}
