% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_text.R
\name{tidy_text}
\alias{tidy_text}
\title{Description}
\usage{
tidy_text(data, column)
}
\arguments{
\item{data}{a tibble produced with the extract_collection() function}

\item{column}{the name of the column to be tokenized}
}
\value{
a tibble
}
\description{
Description
}
\examples{
data_BIOEENVIS=extract_collection("BIOEENVIS", nmax=+Inf)
data_text=tidy_text(data_BIOEENVIS,"title_en")
}
